+++ src/templates/index.js
@@ -10,11 +10,12 @@ const generic = require('./generic');
 const links = require('./links');
 const formatting = require('./formatting');
 const pronounce = require('./pronounce');
+const external = require('./external');
 const ignore = require('./ignore');
 
 //put them all together
 const inlineParsers = Object.assign({}, dates, inline, links, formatting);
-const bigParsers = Object.assign({}, geo, pronounce, misc);
+const bigParsers = Object.assign({}, geo, pronounce, misc, external);
 
 const doTemplate = function(tmpl, wiki, r) {
   let name = getName(tmpl);