+++ tests/tracklist.test.js
@@ -36,8 +36,8 @@ test('track-listing', t => {
 }}`;
   var doc = wtf(str);
   let track = doc.templates(0);
-  t.equal(track.data.headline.text(), 'Side one', 'track-headline');
-  t.equal(track.data.title4.text(), 'Ob-La-Di, Ob-La-Da', 'title4');
+  t.equal(track.data.headline, 'Side one', 'track-headline');
+  t.equal(track.data.title4, 'Ob-La-Di, Ob-La-Da', 'title4');
   t.end();
 });
 
@@ -72,6 +72,6 @@ test('track-listing', t => {
 }}`;
   var doc = wtf(str);
   let track = doc.templates('tracklist')[0];
-  t.equal(track.data.total_length.text(), '23:14', 'track-total_length');
+  t.equal(track.data.total_length, '23:14', 'track-total_length');
   t.end();
 });